index_num and all related values are now 32 bits integers

+++ libsjindex/sj_search_index.h
@@ -17,7 +17,7 @@
 typedef struct {
     uint64_t size; /// size of the input index file
     uint8_t version; /// version number of the index file
-    int64_t index_num; /// number of indexes in the file
+    int index_num; /// number of indexes in the file
     Index *indexes; /// list of indexes read from the file
     char *filename; /// index file name
 } SJ_IndexContext;
